# src/main/swagger/temowl_retrieval.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: "SAM, the semantic aggregation machine"
  name: SAM
  description: >-
              API to retrieve and export Terminologies and their contents.
              Some API Calls expect and/or return a TerminologyID. This ID is configured by the System and depends on the
              system configuration. A list of all available Terminologies can be obtained via the API Call
              `\terminologies`.
              Even though the ID is defined by the system reuse of IDs is intended. For example the TerminologyID of the
              Terminology 'Gene Ontology' should be 'GO'.
              If a Terminology is missing or uses an uncommon TerminologyID please contact the system administrator.
              Issued API Calls with a not present TerminologyID will lead to an error.

              Some API Calls expect and/or return a ConceptID. The ID is either encoded as CURIE
              (https://www.w3.org/TR/2010/NOTE-curie-20101216/) or a canonical IRI (PURL?!).
              A CURIE encoding is used if the originating Terminology (the one that initial defined this Concept) is also present
              in this System, otherwise an IRI is used. However requests issued with an IRI are redirected to a CURIE if
              possible.
              However as multiple IRI could be defined for a single Concept this redirect may not work. Please contact the
...

...

...
              the System.

...
              cases.
...

...

...

...

...


...
basePath: /api/v1
produces:
...
parameters:
    sizeParam:
        in: query
        name: size
        description: Size of pages
        type: integer
        format: int64
        default: 10
    pageParam:
        in: query
        name: page
        description: The requested page
        type: integer
        format: int64
        default: 0
    langParam:
        in: query
        name: lang
        description: This allows to filter the response to only contain descriptions in a
                 specific language. If no description in the given language is present non is returned. The API
                 expect a language to be encoded as defined in https://tools.ietf.org/html/bcp47. Further an
                 extension 'all' is allowed. If that is given properties in all language are returned. The default
                 language is 'en'.
        default: 'en'
        type: string
    versionTagParam:
        in: query
        name: versionTag
        description: Terminology version defined by tag-name
        type: string
...
    terminologyIDParam:
        in: path
        name: terminologyId
        description: Identifier of the terminology.
        required: true
        type: string
...
    autoCompleteQueryParam:
        in: query
        name: q
        description: Beginning of a term to be completed.
        required: true
        type: string
    searchRegexQueryParam:
        in: query
        name: q
        description: Search term.
        required: true
        type: string
    exactMatchParam:
        in: query
        name: exact
        description: Only find exact matches?.
        required: false
        type: boolean
    searchFielsParam:
        in: query
        name: field
        description: To limit results to a property, set field to ID/prefLabel/altLabel/description.
        type: string
        enum:
        - ID
        - prefLabel
        - altLabel
        - description
responses:
    NotFound:
      description: Entity not found.
      schema:
        $ref: '#/definitions/Error'
    BadClient:
      description: Bad Client Request
      schema:
        $ref: '#/definitions/Error'
...
schemes:
...
paths:
...
  '/terminologies/{terminologyId}/concepts/search':
    get:
      tags:
        - retrieval
      summary: Search for concepts
      description: >-
        Returns all search results for your term/REGEX, paginated using page and size. Concept's IDs, names, labels and references are considered. Set exact=true to only recieve perfect matches.
      operationId: Terminology-Seach
      parameters:
        - $ref: '#/parameters/terminologyIDParam'
        - $ref: '#/parameters/versionTagParam'
        - $ref: '#/parameters/langParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/sizeParam'
        - $ref: '#/parameters/searchRegexQueryParam'
        - $ref: '#/parameters/exactMatchParam'
        - $ref: '#/parameters/searchFielsParam'
      responses:
        '200':
          description: Successful response
          schema:
              $ref: '#/definitions/PagedConceptMin'
        '400':
            $ref: '#/responses/BadClient'
        '404':
            $ref: '#/responses/NotFound'
  '/terminologies/{terminologyId}/concepts/autocomplete':
    get:
      tags:
        - retrieval
      summary: Autocomplete concepts
      description: >-
        Returns all possible completions for your term, paginated using page and size. Concept's IDs, names, labels and references are considered.
      operationId: Terminology-Autocompletion
      parameters:
        - $ref: '#/parameters/terminologyIDParam'
        - $ref: '#/parameters/versionTagParam'
        - $ref: '#/parameters/langParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/sizeParam'
        - $ref: '#/parameters/autoCompleteQueryParam'
      responses:
        '200':
          description: Successful response
          schema:
              $ref: '#/definitions/PagedConceptMin'
        '400':
            $ref: '#/responses/BadClient'
        '404':
            $ref: '#/responses/NotFound'
  ########## BEL retrieval
...
definitions:
...
  ConceptMin:
    type: object
    description: "A minimal representation of a Concept. Most information are left out due space constrains."
    required:
      - conceptID
      - link
    properties:
      conceptID:
        description: "A unique ID for the given Concept. This is either a CURIE or an IRI."
        type: string
        example: 'GO:0008289'
      prefLabel:
        description: "The preferred name of a Concept."
        $ref: '#/definitions/Label'
      link:
        description: "This Link can be used to obtain a full representation of the Concept."
        type: string
        example: 'api.example.org//terminologies/GO/concepts/GO:0008289/'
      parent:
        description: "Is the concept a parent of other concepts?"
        type: boolean
        example: true
  Label:
    type: object
    description: "A `name` that describes a Concept."
    required:
      - name
      - lang
    properties:
      name: #ID
        description: "A name that describes a Concept."
        type: string
        example: 'lipid binding'
      lang:
        description: "The language of this Label, as defined by https://tools.ietf.org/html/bcp47."
        type: string
        example: 'en'
...
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
        example: 400
      message:
        type: string
        description: error type
      cause:
        type: string
        description: detailed information on what happened
####################### Pagination extensions
  Page:
    type: object
    required:
      - number
      - numberOfElements
      - totalElements
      - totalPages
      - size
    properties:
        number:
          type: integer
          description: ' Number of this page'
          example: "0"
        numberOfElements:
          type: integer
          description: ' Number of elements in this page'
          example: "4"
        totalElements:
          type: integer
          description: 'total Number of elements'
          example: "4"
        totalPages:
          type: integer
          description: 'total Number of page'
          example: "1"
        size:
          type: integer
          description: 'size of a page'
          example: "10"
  PagedConceptMin:
    allOf:
        - $ref: '#/definitions/Page'
        - type: object
          required:
           - content
          properties:
            content:
                type: array
                items:
                    $ref: '#/definitions/ConceptMin'
...